+++ returns/functions.py
@@ -2,8 +2,6 @@
 
 from typing import Any, Callable, NoReturn, TypeVar
 
-from returns._generated.box import _box as box  # noqa: F401, WPS436
-
 # Aliases:
 _FirstType = TypeVar('_FirstType')
 _SecondType = TypeVar('_SecondType')